+++ internal/serve/httphandler/wallets_handler.go
@@ -51,8 +51,9 @@ func (h *WalletsHandler) GetWallets(w http.ResponseWriter, r *http.Request) {
 func (h *WalletsHandler) parseFilters(ctx context.Context, r *http.Request) ([]data.Filter, error) {
 	filters := []data.Filter{}
 	boolFilterParams := map[string]data.FilterKey{
-		"enabled":      data.FilterEnabledWallets,
-		"user_managed": data.FilterUserManaged,
+		"enabled":         data.FilterEnabledWallets,
+		"user_managed":    data.FilterUserManaged,
+		"include_deleted": data.FilterIncludeDeleted,
 	}
 
 	for param, filterType := range boolFilterParams {
@@ -191,6 +192,10 @@ func (h *WalletsHandler) DeleteWallet(rw http.ResponseWriter, req *http.Request)
 	if err != nil {
 		if errors.Is(err, data.ErrRecordNotFound) {
 			httperror.NotFound("", err, nil).Render(rw)
+			return
+		}
+		if errors.Is(err, data.ErrWalletInUse) {
+			httperror.BadRequest("wallet has pending registrations and cannot be deleted", err, nil).Render(rw)
 			return
 		}
 		httperror.InternalError(ctx, "", err, nil).Render(rw)